refactor add connectionstring builder
