+++ applications/Unity.GrantManager/modules/Unity.TenantManagement/src/Unity.TenantManagement.Application.Contracts/ITenantConnectionStringBuilder.cs
@@ -1,0 +1,9 @@
+﻿using Volo.Abp.Application.Services;
+
+namespace Unity.TenantManagement.Application.Contracts
+{
+    public interface ITenantConnectionStringBuilder : IApplicationService
+    {
+        string Build(string tenantName);
+    }
+}
