+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/Intakes/Index.cshtml
@@ -9,7 +9,7 @@
 @{
     PageLayout.Content.MenuItemName = "GrantManager.Intakes";
     PageLayout.Content.Title = L["Intakes"].Value;    
-    ViewBag.PageTitle = "Intakes";
+    ViewBag.PageTitle = L["Intakes"].Value;
 }
 @section styles {
     <abp-style src="/Pages/Intakes/Index.css" />
